% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{covid19kosovo}
\alias{covid19kosovo}
\title{Get Covi19 Datasets}
\usage{
covid19kosovo(level = c("total", "municipality", "village"))
}
\arguments{
\item{level}{level = c("total", "municipality", "village")}
}
\value{

}
\description{
Get Covi19 Datasets
}
